+++ link-grammar/linkage/score.c
@@ -54,10 +54,10 @@ static int unused_word_cost(Linkage lkg)
  * Computes the cost of the current parse of the current sentence
  * due to the cost of the chosen disjuncts.
  */
-static double compute_disjunct_cost(Linkage lkg)
+static float compute_disjunct_cost(Linkage lkg)
 {
 	size_t i;
-	double lcost;
+	float lcost;
 	lcost =  0.0;
 	for (i = 0; i < lkg->num_words; i++)
 	{
